Split get_frame_allocator out of function.hpp
I'm about to write a new algorithm that injects a frame allocator into
its child's environment, so I'm going to need to be able to refer to
`get_frame_allocator_t` but I don't need to refer to `function` so I'm
splitting the declarations.
